requirements.txt
django
djangorestframework
django-cors-headers
django-environ
psycopg2-binary
pillow
reportlab
gunicorn
dj-database-url
whitenoise
